net: Fix incorrect logging format specifiers
Fix compilation issues that show up if SYS_LOG is mapped to printk
instead of printf. Unlike printf, printk is annotated so that the
compiler catches incorrect format specifiers passed to it.

+++ subsys/net/ip/l2/ethernet.c
@@ -48,10 +48,10 @@ const struct net_eth_addr *net_eth_broadcast_addr(void)
 			 net_sprint_ll_addr(net_nbuf_ll_src(buf)->addr,    \
 					    sizeof(struct net_eth_addr))); \
 									   \
-		NET_DBG("src %s dst %s type 0x%x len %u", out,		   \
+		NET_DBG("src %s dst %s type 0x%x len %zu", out,		   \
 			net_sprint_ll_addr(net_nbuf_ll_dst(buf)->addr,	   \
 					   sizeof(struct net_eth_addr)),   \
-			type, len);					   \
+			type, (size_t)len);				   \
 	} while (0)
 #else
 #define print_ll_addrs(...)
